Crate linera_wasm_instrument

Re-exports

pub use parity_wasm;

Modules

gas_metering

This module is used to instrument a Wasm module with the gas metering code.

Functions

export_mutable_globals

Export all declared mutable globals as prefix_index.

inject_stack_limiter

Inject the instumentation that makes stack overflows deterministic, by introducing an upper bound of the stack size.